--- a/cplugin/certicoq_vanilla_plugin.mlpack
+++ b/cplugin/certicoq_vanilla_plugin.mlpack
@@ -85,7 +85,6 @@ Common0
 Extractable
 Logic0
 
-Char0
 Errors0
 
 FLT
@@ -100,8 +99,6 @@ OrdersTac
 POrderedType
 Specif
 StateMonad
-NPeano
-String1
 Monad_utils
 MSetWeakList
 EqdepFacts