--- a/plugin/_CoqProject
+++ b/plugin/_CoqProject
@@ -92,8 +92,6 @@ extraction/bool.ml
 extraction/bool.mli
 extraction/bracket.ml
 extraction/bracket.mli
-extraction/char0.ml
-extraction/char0.mli
 extraction/byte.mli
 extraction/byte.ml
 extraction/byte0.mli
@@ -304,8 +302,6 @@ extraction/mSetRBT.ml
 extraction/mSetRBT.mli
 extraction/nat0.ml
 extraction/nat0.mli
-extraction/nPeano.ml
-extraction/nPeano.mli
 extraction/number0.ml
 extraction/number0.mli
 extraction/orderedType0.ml
@@ -401,8 +397,6 @@ extraction/stateMonad.ml
 extraction/stateMonad.mli
 extraction/string0.ml
 extraction/string0.mli
-extraction/string1.ml
-extraction/string1.mli
 extraction/templateEnvMap.mli
 extraction/templateEnvMap.ml
 extraction/templateProgram.ml