--- a/plugin/certicoq_plugin.mlpack
+++ b/plugin/certicoq_plugin.mlpack
@@ -77,12 +77,10 @@ EnvMap
 LiftSubst
 UnivSubst0
 WcbvEval
-EtaExpand
 Pretty
 Common0
 Extractable
 Logic0
-TemplateProgram
 
 Char0
 Errors0
@@ -106,6 +104,10 @@ MSetWeakList
 EqdepFacts
 Utils
 Init
+TemplateEnvMap
+TemplateProgram
+EtaExpand
+
 
 Ssreflect
 Ssrbool