--- a/plugin/certicoq_plugin.mlpack
+++ b/plugin/certicoq_plugin.mlpack
@@ -82,7 +82,6 @@ Common0
 Extractable
 Logic0
 
-Char0
 Errors0
 
 FLT
@@ -97,8 +96,6 @@ OrdersTac
 POrderedType
 Specif
 StateMonad
-NPeano
-String1
 Monad_utils
 MSetWeakList
 EqdepFacts