--- a/theories/Extraction/extraction.v
+++ b/theories/Extraction/extraction.v
@@ -159,6 +159,7 @@ Separate Extraction
          cps.M.elements
          Compiler.pipeline.show_IR.
 
+Recursive Extraction Library Ascii.
 Recursive Extraction Library BinPos.
 Recursive Extraction Library OrdersTac.