--- a/theories/ExtractionVanilla/extraction.v
+++ b/theories/ExtractionVanilla/extraction.v
@@ -120,6 +120,7 @@ Separate Extraction
          cps.M.elements
          Compiler.pipeline.show_IR.
 
+Recursive Extraction Library Ascii.
 Recursive Extraction Library BinPos.
 Recursive Extraction Library OrdersTac.