We use lookup maps for a more efficient eta-expansion phase in MetaCoq. This makes the erasure phase drop from 35s to 24s on the certicoq pipeline which goes from template-coq to Clight.

After a bit of profiling, I found that the perf issue in L7 and part of L6 was due to the extlib `nat2string` function which was doing a lot of additions and divisions. Using MetaCoq's `string_of_positive` drops the L7 time on the certicoq_pipeline from 70s to 1s on my computer. The next visible hotspots are the computations of free variables in L6:
it uses a `Module PS := MSetRBT.Make POrderedType.Positive_as_OT.` to keep sets of positive / variable identifiers.
Maybe a good candidate for positive tries?

In the ocaml version of Certicoq, positive is turned into `int` so it is relatively fast, (still gets and sets of maps appear high in the profile), but in the bootstrapped version we have no such luxury yet. In particular the `exp_fv_aux` and `fundefs_fv_aux` functions take a good chunk of the compilation time.
The first profile below is the one of the bootstrapped version, the second is the ocaml one.

This brings down the bootstrapped version to be about 2x slower than the ocaml one, which is not too bad!

![trace](https://user-images.githubusercontent.com/98373/178610614-7a4a88f5-8e2b-4b35-8ee6-090cd4e7a0a0.jpg)

<img width="972" alt="Screenshot 2022-07-13 at 00 48 28" src="https://user-images.githubusercontent.com/98373/178611362-8639e0fd-73c0-46e6-9208-3aa98eea9efe.png">